+++ src/renderer/src/modules/media/media-link-roll/index.ts
@@ -1,0 +1,1 @@
+export { MediaLinkRoll } from './media-link-roll';